slight Refactorings · 0d0b6d9e
